+++ src/Racmacs_wrappers_sera.cpp
@@ -30,6 +30,8 @@ std::string ac_sr_get_name( const AcSerum &sr ){ return sr.get_name(); }
 std::string ac_sr_get_name_full( const AcSerum &sr ){ return sr.get_name_full(); }
 // [[Rcpp::export(rng = false)]]
 std::string ac_sr_get_name_abbreviated( const AcSerum &sr ){ return sr.get_name_abbreviated(); }
+// [[Rcpp::export(rng = false)]]
+std::string ac_sr_get_extra( const AcSerum &sr ){ return sr.get_extra(); }
 
 // [[Rcpp::export(rng = false)]]
 std::string ac_sr_get_match_id( const AcSerum &sr ){ return sr.get_match_id(); }
@@ -57,6 +59,8 @@ AcSerum ac_sr_set_name( AcSerum sr, std::string value ){ sr.set_name(value); ret
 AcSerum ac_sr_set_name_full( AcSerum sr, std::string value ){  sr.set_name_full(value); return sr; }
 // [[Rcpp::export(rng = false)]]
 AcSerum ac_sr_set_name_abbreviated( AcSerum sr, std::string value ){ sr.set_name_abbreviated(value); return sr; }
+// [[Rcpp::export(rng = false)]]
+AcSerum ac_sr_set_extra( AcSerum sr, std::string value ){ sr.set_extra(value); return sr; }
 
 // [[Rcpp::export(rng = false)]]
 AcSerum ac_sr_set_group( AcSerum sr, int value ){ sr.set_group(value); return sr; }